+++ src/main/java/com/jaquadro/minecraft/storagedrawers/block/tile/tiledata/UpgradeData.java
@@ -3,16 +3,14 @@ package com.jaquadro.minecraft.storagedrawers.block.tile.tiledata;
 import com.jaquadro.minecraft.storagedrawers.api.storage.IDrawerAttributesModifiable;
 import com.jaquadro.minecraft.storagedrawers.config.CommonConfig;
 import com.jaquadro.minecraft.storagedrawers.core.ModItems;
-import com.jaquadro.minecraft.storagedrawers.item.EnumUpgradeRedstone;
-import com.jaquadro.minecraft.storagedrawers.item.ItemUpgrade;
-import com.jaquadro.minecraft.storagedrawers.item.ItemUpgradeRedstone;
-import com.jaquadro.minecraft.storagedrawers.item.ItemUpgradeStorage;
+import com.jaquadro.minecraft.storagedrawers.item.*;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.nbt.ListTag;
 import net.minecraft.nbt.Tag;
 import net.minecraft.util.Mth;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.LevelAccessor;
 import org.jetbrains.annotations.NotNull;
 
 import javax.annotation.Nonnull;
@@ -33,6 +31,7 @@ public class UpgradeData extends BlockEntityDataShim
     private boolean hasIllumination;
     private boolean hasFillLevel;
     private boolean hasBalanceFill;
+    private boolean hasRemote;
 
     private IDrawerAttributesModifiable attrs;
 
@@ -115,6 +114,8 @@ public class UpgradeData extends BlockEntityDataShim
             return false;
         if (!(upgrade.getItem() instanceof ItemUpgrade candidate))
             return false;
+        if (!candidate.isEnabled())
+            return false;
 
         if (candidate.getAllowMultiple())
             return true;
@@ -180,6 +181,10 @@ public class UpgradeData extends BlockEntityDataShim
         return hasBalanceFill;
     }
 
+    public boolean hasRemoteUpgrade () {
+        return hasRemote;
+    }
+
     public boolean hasPortabilityUpgrade() {
         for (ItemStack stack : upgrades) {
             if (stack.getItem() == ModItems.PORTABILITY_UPGRADE.get())
@@ -198,6 +203,44 @@ public class UpgradeData extends BlockEntityDataShim
         return -1;
     }
 
+    public ItemStack getRemoteUpgrade () {
+        if (!hasRemote)
+            return null;
+
+        for (ItemStack stack : upgrades) {
+            if (stack.getItem() instanceof ItemUpgradeRemote)
+                return stack;
+        }
+
+        return null;
+    }
+
+    public void unbindRemoteUpgrade () {
+        if (!hasRemote)
+            return;
+
+        for (int i = 0; i < upgrades.length; i++) {
+            ItemStack stack = upgrades[i];
+            if (stack.getItem() instanceof ItemUpgradeRemote remote && remote.isBound()) {
+                upgrades[i] = ItemUpgradeRemote.setUnbound(stack);
+                onUpgradeChanged(stack, upgrades[i]);
+            }
+        }
+    }
+
+    public void updateRemoteUpgradeBinding (ItemStack refStack) {
+        if (!hasRemote || refStack == null)
+            return;
+
+        for (int i = 0; i < upgrades.length; i++) {
+            ItemStack stack = upgrades[i];
+            if (stack.getItem() instanceof ItemUpgradeRemote) {
+                upgrades[i] = ItemUpgradeRemote.copyControllerBinding(refStack, upgrades[i]);
+                onUpgradeChanged(stack, upgrades[i]);
+            }
+        }
+    }
+
     private void syncUpgrades () {
         if (this.attrs == null)
             return;
@@ -213,6 +256,7 @@ public class UpgradeData extends BlockEntityDataShim
         hasIllumination = false;
         hasFillLevel = false;
         hasBalanceFill = false;
+        hasRemote = false;
 
         for (ItemStack stack : upgrades) {
             Item item = stack.getItem();
@@ -220,7 +264,7 @@ public class UpgradeData extends BlockEntityDataShim
             if (item == ModItems.ONE_STACK_UPGRADE.get())
                 hasOneStack = true;
             else if (item == ModItems.VOID_UPGRADE.get())
-                hasVoid = true;
+                hasVoid = CommonConfig.UPGRADES.enableVoidUpgrade.get();
             //else if (item == ModItems.CONVERSION_UPGRADE.get())
             //    hasConversion = true;
             else if (item == ModItems.CREATIVE_STORAGE_UPGRADE.get())
@@ -228,11 +272,16 @@ public class UpgradeData extends BlockEntityDataShim
             else if (item == ModItems.CREATIVE_VENDING_UPGRADE.get())
                 hasVending = true;
             else if (item == ModItems.ILLUMINATION_UPGRADE.get())
-                hasIllumination = true;
+                hasIllumination = CommonConfig.UPGRADES.enableIlluminationUpgrade.get();
             else if (item == ModItems.FILL_LEVEL_UPGRADE.get())
-                hasFillLevel = true;
+                hasFillLevel = CommonConfig.UPGRADES.enableFillLevelUpgrade.get();
             else if (item == ModItems.BALANCE_FILL_UPGRADE.get())
-                hasBalanceFill = true;
+                hasBalanceFill = CommonConfig.UPGRADES.enableBalanceUpgrade.get();
+            else if (item instanceof ItemUpgradeRemote remote) {
+                hasRemote = remote.isGroupUpgrade()
+                    ? CommonConfig.UPGRADES.enableRemoteGroupUpgrade.get()
+                    : CommonConfig.UPGRADES.enableRemoteUpgrade.get();
+            }
         }
 
         attrs.setIsVoid(hasVoid);