+++ src/main/java/com/jaquadro/minecraft/storagedrawers/core/PlayerEventListener.java
@@ -3,6 +3,9 @@ package com.jaquadro.minecraft.storagedrawers.core;
 import com.jaquadro.minecraft.storagedrawers.api.storage.attribute.IPortable;
 import com.jaquadro.minecraft.storagedrawers.config.CommonConfig;
 
+import com.jaquadro.minecraft.storagedrawers.inventory.ContainerDrawers;
+import com.jaquadro.minecraft.storagedrawers.item.ItemUpgradeRemote;
+import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.effect.MobEffectInstance;
 import net.minecraft.world.effect.MobEffects;
 import net.minecraft.world.entity.player.Inventory;
@@ -12,7 +15,9 @@ import net.minecraft.world.item.ItemStack;
 import net.minecraftforge.event.TickEvent.Phase;
 import net.minecraftforge.event.TickEvent.PlayerTickEvent;
 import net.minecraftforge.event.entity.player.EntityItemPickupEvent;
+import net.minecraftforge.event.entity.player.PlayerContainerEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
+import net.minecraftforge.fml.LogicalSide;
 
 /** Punishes players holding filled drawers, if enabled in config */
 public class PlayerEventListener {
@@ -37,6 +42,9 @@ public class PlayerEventListener {
 		if(event.phase != Phase.END || event.player.tickCount % 60 != 0)
 			return;
 
+		if (event.side == LogicalSide.SERVER)
+			ItemUpgradeRemote.validateInventory(event.player.getInventory(), event.player.level());
+
 		if (!CommonConfig.GENERAL.heavyDrawers.get())
 			return;
 
@@ -51,6 +59,16 @@ public class PlayerEventListener {
 				return;
 		}
 	}
+
+	/*@SubscribeEvent
+	public void onPlayerContainerOpen (PlayerContainerEvent.Open event) {
+		if (!(event.getEntity() instanceof ServerPlayer))
+			return;
+
+		ItemUpgradeRemote.validateInventory(event.getEntity().getInventory(), event.getEntity().level());
+		//if (event.getContainer() instanceof ContainerDrawers)
+		//	ItemUpgradeRemote.validateInventory(event.getContainer()., event.getEntity().level());
+	}*/
 
 	private boolean checkItemDebuf (ItemStack stack, Player player) {
 		Item item = stack.getItem();